refactor(plugin/antd): [UserMenu] extract to new file
