+++ plugins/antd/src/components/layout/LeftSideMenuTopHeaderLayout/Header/Header.tsx
@@ -1,59 +1,17 @@
-import { coreActions, CoreStore, Logo, useAuth } from '@onr/core';
-import { Avatar, Layout, Menu, MenuProps } from 'antd';
+import { coreActions, CoreStore, Logo } from '@onr/core';
+import { Layout } from 'antd';
 import Link from 'next/link';
 import { ComponentType, FC, ReactNode } from 'react';
 import { FiBarChart } from 'react-icons/fi';
 import { useDispatch, useSelector } from 'react-redux';
+import { UserMenu } from '../../components';
 import { DashHeader } from '../../HeaderBarLeftSideMenuLayout/Header';
 
-interface AvatarProps {
+interface Props {
+  HeaderMainSection?: ComponentType;
+  logo?: ReactNode;
   avatar?: string;
 }
-
-interface LogoProps {
-  logo?: ReactNode;
-}
-
-interface Props extends AvatarProps, LogoProps {
-  HeaderMainSection?: ComponentType;
-}
-
-const UserMenu: FC<AvatarProps> = ({ avatar }) => {
-  const dispatch = useDispatch();
-  const { user, signOut, signIn } = useAuth();
-
-  const items: MenuProps['items'] = [
-    {
-      key: 'submenu',
-      icon: <Avatar src={avatar ?? '/static/images/avatar.jpg'} />,
-      children: [
-        {
-          key: 'settings',
-          label: 'Settings',
-          onClick: () => dispatch(coreActions.setOptionDrawer()),
-        },
-        {
-          label: 'Profile',
-          key: 'profile',
-        },
-        {
-          type: 'divider',
-        },
-        {
-          label: (
-            <>
-              {user && <a onClick={signOut}>Sign Out</a>}
-              {!user && <a onClick={() => signIn()}>Sign In</a>}
-            </>
-          ),
-          key: 'auth',
-        },
-      ],
-    },
-  ];
-
-  return <Menu mode="horizontal" items={items} />;
-};
 
 export const Header: FC<Props> = props => {
   const { HeaderMainSection, avatar, logo } = props;
