+++ plugins/antd/src/components/layout/components/index.ts
@@ -1,0 +1,1 @@
+export * from './UserMenu';
